feat: reservation crud

+++ src/EcoPark.Application/Reservations/Delete/DeleteReservationCommandHandler.cs
@@ -9,24 +9,30 @@ public class DeleteReservationCommandHandler(IRepository<ReservationModel> repos
 
         try
         {
-            await repository.UnitOfWork.StartAsync(cancellationToken);
-
-            var databaseOperationResult = await repository.DeleteAsync(command, cancellationToken);
-
-            if (databaseOperationResult)
+            if (await repository.CheckChangePermissionAsync(command, cancellationToken))
             {
-                await repository.UnitOfWork.SaveEntitiesAsync(cancellationToken);
-                await repository.UnitOfWork.CommitAsync(cancellationToken);
+                await repository.UnitOfWork.StartAsync(cancellationToken);
 
-                result = new DatabaseOperationResponseViewModel("Delete", EOperationStatus.Successful,
-                    "Reservation was deleted successfully!");
+                var databaseOperationResult = await repository.DeleteAsync(command, cancellationToken);
+
+                if (databaseOperationResult)
+                {
+                    await repository.UnitOfWork.SaveEntitiesAsync(cancellationToken);
+                    await repository.UnitOfWork.CommitAsync(cancellationToken);
+
+                    result = new DatabaseOperationResponseViewModel("Delete", EOperationStatus.Successful,
+                        "Reservation was deleted successfully!");
+                }
+                else
+                {
+                    await repository.UnitOfWork.RollbackAsync(cancellationToken);
+                    result = new DatabaseOperationResponseViewModel("Delete", EOperationStatus.Failed,
+                        "No Reservations were found with this id");
+                }
             }
             else
-            {
-                await repository.UnitOfWork.RollbackAsync(cancellationToken);
-                result = new DatabaseOperationResponseViewModel("Delete", EOperationStatus.Failed,
-                    "No Reservations were found with this id");
-            }
+                result = new("Delete", EOperationStatus.NotAuthorized,
+                    "You have no permission to delete this reservation");
         }
         catch (Exception e)
         {
